i18n code moved into services
