+++ apps/next-client/src/app/[locale]/layout.tsx
@@ -7,7 +7,7 @@ import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter';
 import { ThemeProvider } from '@mui/material/styles';
 import theme from '@/assets/styles/theme';
 import { AppBar, MixPanelClient, PageContainer } from '@/components';
-import { routing } from '@/i18n/routing';
+import { i18nRouting } from '@/services/i18n';
 import { type Locales } from '@/types';
 import '../globals.css';
 
@@ -42,7 +42,7 @@ export default async function RootLayout({
 }: RootLayoutProps) {
   const { locale } = await params;
   /* Ensure that the incoming `locale` is valid */
-  if (!routing.locales.includes(locale)) {
+  if (!i18nRouting.locales.includes(locale)) {
     notFound();
   }
