+++ apps/next-client/src/components/Link.tsx
@@ -1,7 +1,7 @@
 import { type ReactNode, type ComponentProps } from 'react';
 import Button from '@mui/material/Button';
 import ArrowOutwardIcon from '@mui/icons-material/ArrowOutward';
-import { Link } from '@/i18n/routing';
+import { Link } from '@/services/i18n';
 
 /**
  * The Link component from i18n mostly extends the same component
