+++ apps/next-client/src/middleware.ts
@@ -1,12 +1,12 @@
 import createMiddleware from 'next-intl/middleware';
-import { routing } from '@/i18n/routing';
+import { i18nRouting } from '@/services/i18n';
 
 /**
  * For using next-intl along with other middlewares,
  * https://next-intl.dev/docs/routing/middleware#composing-other-middlewares
  */
 
-export default createMiddleware(routing);
+export default createMiddleware(i18nRouting);
 
 export const config = {
   // Match only internationalized pathnames
@@ -16,6 +16,6 @@ export const config = {
     // Match all pathnames except for
     // - if they start with `/api`, `/_next` or `/_vercel`
     // - the ones containing a dot (e.g. `favicon.ico`)
-    '/((?!api|_next|_vercel|.*\\..*).*)',
+    '/((?!api|_next|_vercel|.*\\..*).*)'
   ]
 };
