+++ apps/next-client/src/services/i18n/request.ts
@@ -2,7 +2,7 @@
  * https://next-intl.dev/docs/getting-started/app-router/with-i18n-routing#i18n-request-path
  */
 import { getRequestConfig } from 'next-intl/server';
-import { routing } from './routing';
+import { i18nRouting } from './routing';
 import { type Locales } from '@/types';
 
 export default getRequestConfig(async ({ requestLocale }) => {
@@ -20,8 +20,8 @@ export default getRequestConfig(async ({ requestLocale }) => {
    * Extract "en" from "en-US"
    * const preferredLang = navigator.language.split('-')[0];
    */
-  if (!locale || !routing.locales.includes(locale as Locales)) {
-    locale = routing.defaultLocale;
+  if (!locale || !i18nRouting.locales.includes(locale as Locales)) {
+    locale = i18nRouting.defaultLocale;
   }
 
   return {
@@ -32,6 +32,6 @@ export default getRequestConfig(async ({ requestLocale }) => {
      * but it is recommended to define all translations in code, for
      * type checking.
      */
-    messages: (await import(`@/languages/${locale}.json`)).default
+    messages: (await import(`@/services/i18n/languages/${locale}.json`)).default
   };
 });
