+++ apps/next-client/src/services/i18n/routing.ts
@@ -2,7 +2,7 @@ import { defineRouting } from 'next-intl/routing';
 import { createNavigation } from 'next-intl/navigation';
 import { Locales } from '@/types';
 
-export const routing = defineRouting({
+export const i18nRouting = defineRouting({
   // A list of all locales that are supported
   locales: Object.values(Locales),
 
@@ -16,4 +16,4 @@ export const routing = defineRouting({
 // Lightweight wrappers around Next.js' navigation APIs
 // that will consider the routing configuration
 export const { Link, redirect, usePathname, useRouter, getPathname }
-  = createNavigation(routing);
+  = createNavigation(i18nRouting);
